requirements.txt

click>=8.1.0
pyembroidery>=1.4.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
